paper_implementation/requirements.txt
numpy>=1.21.0
scipy>=1.7.0
Pillow>=8.0.0



